+++ agate/table.py
@@ -308,7 +308,7 @@ class Table(utils.Patchable):
             return self._fork(self.rows, column_names, self._column_types, row_names=row_names)
 
     @classmethod
-    def from_csv(cls, path, column_names=None, column_types=None, row_names=None, header=True, snifflimit=0, **kwargs):
+    def from_csv(cls, path, column_names=None, column_types=None, row_names=None, header=True, sniff_limit=0, **kwargs):
         """
         Create a new table for a CSV. This method uses agate's builtin
         CSV reader, which supports unicode on both Python 2 and Python 3.
@@ -328,7 +328,7 @@ class Table(utils.Patchable):
             and will be skipped. If `header` and `column_names` are both
             specified then a row will be skipped, but `column_names` will be
             used.
-        :param snifflimit:
+        :param sniff_limit:
             Limit CSV dialect sniffing to the specified number of bytes. Set to 
             None to sniff the entire file. Defaults to 0 or no sniffing.
         """
@@ -338,10 +338,10 @@ class Table(utils.Patchable):
             with open(path) as f:
                 contents = f.read()
         
-        if snifflimit is None:
+        if sniff_limit is None:
             kwargs['dialect'] = csv.Sniffer().sniff(contents)
-        elif snifflimit > 0:
-            kwargs['dialect'] = csv.Sniffer().sniff(contents[:snifflimit])
+        elif sniff_limit > 0:
+            kwargs['dialect'] = csv.Sniffer().sniff(contents[:sniff_limit])
         
         rows = list(csv.reader(StringIO(contents), **kwargs))